refactor : 뷰 로직 분리 , handle 함수들 useManageUser로 분리
