+++ src/1_Page/Setting/model/useTab.js
@@ -4,7 +4,7 @@ const useTab = () => {
   const [activeTab, setActiveTab] = useState("화이트");
   const handleTabWhite = () => setActiveTab("화이트");
   const handleTabDark = () => setActiveTab("다크");
-  const handleGetPresentTab = (selectTab) => activeTab === selectTab;
-  return { activeTab, handleTabWhite, handleTabDark, handleGetPresentTab };
+  const isPresentTab = (selectTab) => activeTab === selectTab;
+  return { activeTab, handleTabWhite, handleTabDark, isPresentTab };
 };
 export default useTab;
